requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
dj-database-url==2.0.0
Django==4.2.1
django-heroku==0.3.1
djangorestframework==3.14.0
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
idna==3.4
psycopg2==2.9.6
pytz==2023.3
requests==2.30.0
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
whitenoise==6.4.0
twilio==8.4.0